source/requirements.txt
ibm_db
flask
requests
Pillow
bcrypt
